long long

+++ 14.cc
@@ -87,7 +87,7 @@ int main() {
   collatzSequence(1);
   longestSequence[1] = 1;
   for (long i = 2; i <= maxTestValue; i++) {
-    long thisSequenceLength = collatzSequence(i);
+    long long thisSequenceLength = collatzSequence(i);
     long long previousLargest = longestSequence[i-1];
     if (sequenceLengths[previousLargest] > sequenceLengths[i]) {
       longestSequence[i] = previousLargest;
